{"schema":"decodex/commit/1","summary":"Split Decodex app account row view","authority":"XY-1115"}
